Everything is almost functional
I can't propagate bundleIsServer et. al to the other scripts tho :(
