+++ package.json
@@ -6,12 +6,11 @@
     "lint": "scripts/lint_and_checktypes.sh",
     "acceptmigrations": "ts-node scripts/acceptMigrations.ts || git diff --no-index schema/accepted_schema.sql schema/schema_to_accept.sql",
     "makemigrations": "ts-node scripts/makeMigrations.ts",
-    "tsc": "tsc --noEmit --skipLibCheck",
     "eslint": "eslint --max-warnings 0 --cache --ext .jsx,js,tsx,ts packages",
     "build": "ts-node ./build.ts",
     "start": "pnpm build -r",
     "prod": "NODE_ENV=production pnpm start",
-    "migrate": "node ./migrate.js",
+    "migrate": "ts-node ./migrate.ts",
     "branchdb": "ts-node scripts/branchDb.ts",
     "dev": "pnpm build --watch --run",
     "start-dev-db": "pnpm build --no-clear -run --watch --lint --db ../LessWrong-Credentials/connectionConfigs/dev.json --settings ../LessWrong-Credentials/settings-local-dev-devdb.json",
