+++ packages/lesswrong/lib/types/globals.d.ts
@@ -1,0 +1,18 @@
+// declare var bundleIsServer: boolean;
+// declare var bundleIsTest: boolean;
+// declare var bundleIsProduction: boolean;
+// declare var bundleIsMigrations: boolean;
+// declare var defaultSiteAbsoluteUrl: string;
+// declare var serverPort: number;
+// declare var estrellaPid: number;
+declare global {
+  var bundleIsServer: boolean;
+  var bundleIsTest: boolean;
+  var bundleIsProduction: boolean;
+  var bundleIsMigrations: boolean;
+  var defaultSiteAbsoluteUrl: string;
+  var serverPort: number;
+  var estrellaPid: number;
+}
+
+export {};
