+++ packages/lesswrong/migration_env.ts
@@ -1,0 +1,63 @@
+import "./server/vulcan-lib/index";
+import "./server/apolloServer";
+import "./lib/index";
+import "./server/tagging/tagCallbacks";
+import "./server/tagging/tagsGraphQL";
+
+import "./server/callbacks/commentCallbacks";
+import "./server/callbacks/conversationCallbacks";
+import "./server/callbacks/localgroupCallbacks";
+import "./server/callbacks/gardenCodeCallbacks";
+import "./server/resolvers/commentResolvers";
+import "./server/resolvers/notificationResolvers";
+import "./server/resolvers/dialogueMessageResolvers";
+import "./server/callbacks/postCallbacks";
+import "./server/posts/validatePost";
+import "./server/callbacks/chapterCallbacks";
+import "./server/callbacks/sequenceCallbacks";
+import "./server/callbacks/bookCallbacks";
+import "./server/callbacks/collectionCallbacks";
+import "./server/callbacks/messageCallbacks";
+import "./server/callbacks/revisionCallbacks";
+import "./server/callbacks/userCallbacks";
+import "./server/callbacks/tagRelsCallbacks";
+// import "./server/staticRoutes/debugHeaders";
+// import "./server/tableOfContents";
+import "./server/callbacks/subscriptionCallbacks";
+import "./server/callbacks/rateLimitCallbacks";
+import "./server/callbacks/reviewVoteCallbacks";
+import "./server/callbacks/tagFlagCallbacks";
+import "./server/callbacks/moderatorActionCallbacks";
+import "./server/callbacks/advisorRequestsCallbacks";
+import "./server/callbacks/digestCallbacks";
+import "./server/callbacks/dialogueMatchPreferenceCallbacks";
+
+import "./server/resolvers/alignmentForumMutations";
+import "./server/callbacks/alignment-forum/callbacks";
+import "./server/callbacks/alignment-forum/alignmentCommentCallbacks";
+import "./server/callbacks/alignment-forum/alignmentPostCallbacks";
+import "./server/callbacks/alignment-forum/alignmentUserCallbacks";
+import "./server/callbacks/votingCallbacks";
+
+import "./server/resolvers/diffResolvers";
+import "./server/resolvers/revisionResolvers";
+import "./server/resolvers/postResolvers";
+import "./server/resolvers/userResolvers";
+import "./server/resolvers/karmaChangeResolvers";
+import "./server/resolvers/coronaLinkDatabase";
+import "./server/resolvers/mozillaHubsData";
+import "./server/resolvers/arbitalPageData";
+import "./server/resolvers/tagHistoryFeed";
+import "./server/resolvers/tagResolvers";
+import "./server/resolvers/allTagsActivityFeed";
+import "./server/resolvers/recentDiscussionFeed";
+import "./server/resolvers/elicitPredictions";
+import "./server/resolvers/reviewVoteResolvers";
+import "./server/resolvers/petrovDayResolvers";
+import "./server/resolvers/analyticsResolvers";
+import "./server/resolvers/moderationResolvers";
+import "./server/resolvers/typingIndicatorsResolvers";
+import "./server/resolvers/dialogueChecksResolvers";
+import "./server/resolvers/electionCandidateResolvers";
+import "./server/resolvers/electionVoteResolvers";
+import "./server/resolvers/givingSeasonResolvers";
