+++ tsconfig.json
@@ -19,7 +19,7 @@
     "noEmit": true
   },
   "include": ["packages/lesswrong", "public/lesswrong-editor/src", "cypress"],
-  "files": ["./build.ts"],
+  "files": ["./build.ts", "./migrate.ts"],
   "exclude": ["node_modules", "build"],
   "ts-node": {
     "files": true
